Fix of warning when using sqlite with SQLITE_HAS_CODEC

+++ src/FMDatabase.m
@@ -178,7 +178,7 @@
         return NO;
     }
     
-    int rc = sqlite3_rekey(db, [key UTF8String], strlen([key UTF8String]));
+    int rc = sqlite3_rekey(db, [key UTF8String], (int)strlen([key UTF8String]));
     
     if (rc != SQLITE_OK) {
         NSLog(@"error on rekey: %d", rc);
@@ -197,7 +197,7 @@
         return NO;
     }
     
-    int rc = sqlite3_key(db, [key UTF8String], strlen([key UTF8String]));
+    int rc = sqlite3_key(db, [key UTF8String], (int)strlen([key UTF8String]));
     
     return (rc == SQLITE_OK);
 #else
